[#24 #88 #92 #102] Heuristik angepasst, RandomAI weint
* TreeNode: float Werte gegen Doubles ausgetauscht, ist zwar schlechter
für die Performance aber die Ergebnisse sind genauer, es haben nicht
mehr mehrere Knoten den selben UCT wert
* MCTS/Constants: In der Heuristik die Euklidische Distanz aller Pieces
aller Teams zu den Bases der Gegnern eingefügt, jetzt wollen die Pieces
auch wirklich in die Base laufen
* Test Klasse: An Änderungen angepasst und Spielsimulation verbessert

+++ ctf-ai/src/main/java/org/ctf/ai/mcts/TreeNode.java
@@ -44,17 +44,17 @@ public class TreeNode implements Comparable<TreeNode> {
     * due to the creation of children the player is stored as a boolean attribute of this class.
     * @return V value for UCT
     */
-   public float getV() {
+   public double getV() {
      int team = (gameState.getCurrentTeam()-1) % gameState.getTeams().length;
      team = team >= 0 ? team : team*-1;
-     return wins[team] / (float)getNK();
+     return wins[team] / (double)getNK();
    }
 
    /**
     * @return returns the UCT value of the current node
     */
-   public float getUCT(float C) {
-     return getV() + C * (float)Math.sqrt((float)Math.log(parent.getNK()) / getNK());
+   public double getUCT(double C) {
+     return getV() + C * Math.sqrt((double)Math.log(parent.getNK()) / getNK());
    }
 
    /**
@@ -104,7 +104,7 @@ public class TreeNode implements Comparable<TreeNode> {
 
    @Override
    public int compareTo(TreeNode node) {
-     return Float.compare(node.getV(), getV());
+     return Double.compare(node.getV(), getV());
    }
 
    /**
